+++ let-engine/src/prelude.rs
@@ -8,33 +8,32 @@
 
 // Resources only exists if client is enabled.
 
+#[cfg(feature = "client")]
 pub use let_engine_core::resources::*;
 
 pub use let_engine_core::{camera::*, objects::*};
 
 pub use crate::*;
+#[cfg(feature = "client")]
 pub use data::*;
 
 // Client structs
 #[cfg(feature = "client")]
-pub use crate::events::*;
+mod client {
+    pub use super::materials::*;
+    pub use super::textures::*;
+    pub use super::window::*;
+    pub use crate::events::*;
+    pub use let_engine_core::draw::PresentMode;
+}
 #[cfg(feature = "client")]
-pub use let_engine_core::draw::PresentMode;
-#[cfg(feature = "client")]
-pub use materials::*;
-#[cfg(feature = "client")]
-pub use textures::*;
-#[cfg(feature = "client")]
-pub use window::*;
+pub use client::*;
 
 // Physics structs
 #[cfg(feature = "physics")]
 pub use joints::*;
 #[cfg(feature = "physics")]
 pub use physics::*;
-
-// Label structs
-pub use let_engine_widgets::labels::*;
 
 // Audio structs
 #[cfg(feature = "audio")]
@@ -43,6 +42,10 @@ pub use let_engine_audio::*;
 // Asset system
 #[cfg(feature = "asset_system")]
 pub use asset_system::*;
+
+// Networking
+#[cfg(feature = "networking")]
+pub use networking::*;
 
 // Other structs
 pub use crate::settings::{EngineSettings, EngineSettingsBuilder, EngineSettingsBuilderError};