+++ test.sh
@@ -1,6 +1,6 @@
+cargo clippy --no-default-features
 cargo clippy
 cargo clippy --all-features
-cargo clippy --no-default-features
+cargo test --no-default-features
 cargo test
 cargo test --all-features
-cargo test --no-default-features